feat(react): add host rspack module federation support
