+++ packages/react/src/generators/host/files/rspack-module-federation-ssr/module-federation.server.config.js__tmpl__
@@ -1,0 +1,16 @@
+// @ts-check
+
+/**
+ * @type {import('@nx/rspack/module-federation').ModuleFederationConfig}
+ **/
+const moduleFederationConfig = {
+    name: '<%= projectName %>',
+    remotes: [
+       <% if (static) { 
+         remotes.forEach(function(r) { %> "<%= r.fileName %>", <% }); 
+       }
+    %>
+  ],
+};
+
+module.exports = moduleFederationConfig;
