+++ packages/react/src/generators/host/files/webpack-module-federation-ssr-ts/server.ts__tmpl__
