+++ packages/react/src/generators/host/files/webpack-module-federation-ssr/module-federation.server.config.js__tmpl__
