+++ packages/react/src/generators/host/files/webpack-module-federation-ssr/server.ts__tmpl__
