+++ packages/react/src/generators/host/files/webpack-module-federation-ssr/tsconfig.server.json__tmpl__
