+++ packages/react/src/generators/host/files/webpack-module-federation-ssr/webpack.server.config.js__tmpl__
