+++ packages/react/src/generators/host/files/webpack-module-federation-ts/module-federation.config.ts__tmpl__
