+++ packages/react/src/generators/host/files/webpack-module-federation-ts/webpack.config.prod.ts__tmpl__
