+++ packages/react/src/generators/host/files/webpack-module-federation-ts/webpack.config.ts__tmpl__
