+++ packages/react/src/generators/host/files/webpack-module-federation/module-federation.config.js__tmpl__
