+++ packages/react/src/generators/host/files/webpack-module-federation/webpack.config.js__tmpl__
