+++ packages/react/src/generators/host/files/webpack-module-federation/webpack.config.prod.js__tmpl__
